feat(roomsservice):
se agruego la clase roomsservise

+++ src/main/java/com/anonymous/zanithresort/repository/RoomsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.anonymous.zanithresort.model.Rooms;
 
-public interface RoomsRepository  extends JpaRepository<Rooms, Integer>{
+public interface RoomsRepository  extends JpaRepository<Rooms, Long>{
 
 }
